server/requirements.txt
python-socketio
eventlet
firebase-admin